Update agent location on LocationScreen resume
If passengers are ejected from a vehicle due to changes made in the
VEquipScreen and the agent list is not updated the agents appear to
disappear until the player moves the mouse over the agent list and
triggers an update. This commit tries to avoid that by forcing an
update whenever the player returns to the LocationScreen
from another form.